chore: Moves the hooks folder to the src folder
