+++ superset-frontend/src/components/FaveStar/index.tsx
@@ -20,7 +20,7 @@
 import React, { useCallback } from 'react';
 import { t, styled } from '@superset-ui/core';
 import { Tooltip } from 'src/components/Tooltip';
-import { useComponentDidMount } from 'src/common/hooks/useComponentDidMount';
+import { useComponentDidMount } from 'src/hooks/useComponentDidMount';
 import Icons from 'src/components/Icons';
 
 interface FaveStarProps {
