+++ superset-frontend/src/components/OmniContainer/index.tsx
@@ -21,7 +21,7 @@ import React, { useRef, useState } from 'react';
 import { styled, t } from '@superset-ui/core';
 import { isFeatureEnabled, FeatureFlag } from 'src/featureFlags';
 import Modal from 'src/components/Modal';
-import { useComponentDidMount } from 'src/common/hooks/useComponentDidMount';
+import { useComponentDidMount } from 'src/hooks/useComponentDidMount';
 import { logEvent } from 'src/logger/actions';
 import { Omnibar } from './Omnibar';
 import { LOG_ACTIONS_OMNIBAR_TRIGGERED } from '../../logger/LogUtils';
