+++ superset-frontend/src/dashboard/components/DashboardBuilder/DashboardBuilder.tsx
@@ -34,7 +34,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { getUrlParam } from 'src/utils/urlUtils';
 import { DashboardLayout, RootState } from 'src/dashboard/types';
 import { setDirectPathToChild } from 'src/dashboard/actions/dashboardState';
-import { useElementOnScreen } from 'src/common/hooks/useElementOnScreen';
+import { useElementOnScreen } from 'src/hooks/useElementOnScreen';
 import { FeatureFlag, isFeatureEnabled } from 'src/featureFlags';
 import {
   deleteTopLevelTabs,
