+++ superset-frontend/src/dashboard/components/menu/ShareMenuItems/index.tsx
@@ -17,7 +17,7 @@
  * under the License.
  */
 import React from 'react';
-import { useUrlShortener } from 'src/common/hooks/useUrlShortener';
+import { useUrlShortener } from 'src/hooks/useUrlShortener';
 import copyTextToClipboard from 'src/utils/copy';
 import { t } from '@superset-ui/core';
 import { Menu } from 'src/common/components';
